requirements.txt
open3d==0.18.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
openai>=1.12.0
python-dotenv>=1.0.0
tqdm>=4.66.0